requirements.txt
numpy<2
pandas
pyarrow
datasets>=2.0.0
apache-beam>=2.40.0
spacy>=3.0.0
networkx>=2.6.0
transformers>=4.0.0
gradio>=3.0.0
sentence-transformers
faiss-cpu
ollama
scipy
